ai-engine/pyproject.toml
[project]
name = "ai-engine"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.128.0",
    "numpy>=2.4.2",
    "pydantic>=2.12.5",
    "pypdf2>=3.0.1",
    "python-dotenv>=1.2.1",
    "python-multipart>=0.0.22",
    "sentence-transformers>=5.2.2",
    "supabase>=2.27.3",
    "uvicorn[standard]>=0.40.0",
]
